addition of the game method and correction in the code
addition of the game method which itself makes two people play, it manages the different type of error when a user enters the column number and displays the game board as the game progresses. game and at the end by displaying the winner. 

modification of the return of the check winner method so that it only returns an enumeration of Token. removal of an unreachable pattern in the display method. correction of the is_full method.

validation of another objective in README.md