addition of launch_rounds function and launch_games_thread function
added launch_round function which starts a certain number of games and returns each player's winning taut and draw rate in an array. addition of the launch_games_thread function which launches a certain number of threads of the launch_round function then displays the execution time and the various win or draw rates. Removal of the associated function game being useless in the branch. modification of the rand_game function to be able to choose if the first player is chosen randomly or not. validation of several objectives in the roadmap in README.md